backend/requirements.txt
asgiref==3.11.0
asttokens==3.0.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
executing==2.2.1
gunicorn==25.1.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
six==1.17.0
sqlparse==0.5.4
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.14
